update nitrite to 4.2.1.

+++ chemical_db_oss/chemdb_sql_jdbi_json_oss/src/main/java/de/unijena/bioinf/chemdb/SpectralSqlJdbiJsonDatabase.java
@@ -55,7 +55,7 @@ public class SpectralSqlJdbiJsonDatabase implements SpectralLibrary {
     }
 
     @Override
-    public int countAllSpectra() {
+    public long countAllSpectra() {
         return jdbi.withExtension(Dao.class, Dao::countAll);
     }
